+++ docs/repl/ReplPlayground.vue
@@ -34,6 +34,10 @@ const showReloadMsg = computed(() => {
   )
 })
 
+async function doLoadPg() {
+  pg.value = await loadPg()
+}
+
 async function loadPg() {
   const extensions = Object.fromEntries(
     enabledExtensions.value.map((extension) => {
@@ -45,14 +49,51 @@ async function loadPg() {
   )
 
   loadedExtensions.value = [...enabledExtensions.value]
-  pg.value = await PGlite.create({
-    dataDir: 'idb://pglite-playground',
+  const dbName = 'pglite-playground'
+  try {
+    console.log(`Creating PGlite instance with idb://${dbName}`)
+    return await PGlite.create({
+      dataDir: `idb://${dbName}`,
+      extensions,
+    })
+  } catch {
+    console.error(`Failed to create PGlite instance on idb at ${dbName}`)
+  }
+  // Delete the IndexedDB for '/pglite-playground' before creating PGlite instance again
+  if (window.indexedDB) {
+    try {
+      console.log('Deleting IndexedDB for', dbName)
+      await new Promise((resolve, reject) => {
+        const req = window.indexedDB.deleteDatabase(`/pglite/${dbName}`)
+        req.onsuccess = resolve
+        req.onerror = reject
+        req.onblocked = resolve // In case user blocks the deletion
+      })
+    } catch (err) {
+      console.error(`Failed to delete ${dbName} IndexedDB:`, err)
+    }
+    try {
+      return await PGlite.create({
+        dataDir: `idb://${dbName}`,
+        extensions,
+      })
+    } catch (err) {
+      console.error(
+        `Failed to create PGlite with idb://${dbName} instance again after trying to delete it`,
+        err,
+      )
+    }
+  } else {
+    console.error('IndexedDB is not supported')
+  }
+  console.log('Creating an in-memory db')
+  return await PGlite.create({
     extensions,
   })
 }
 
-onMounted(() => {
-  loadPg()
+onMounted(async () => {
+  pg.value = await loadPg()
 })
 
 const rootStyle = window.getComputedStyle(document.body)
@@ -138,7 +179,7 @@ async function clearDb() {
     if (closed) break
     await new Promise((resolve) => setTimeout(resolve, 10))
   }
-  await loadPg()
+  pg.value = await loadPg()
 }
 </script>
 
@@ -184,7 +225,7 @@ async function clearDb() {
     </div>
     <div class="main" @click="showOptions = false">
       <div class="info-msg" v-if="showReloadMsg">
-        Please <button @click="loadPg()">Restart</button> PGlite to enable the
+        Please <button @click="doLoadPg()">Restart</button> PGlite to enable the
         selected extensions.
       </div>
       <pglite-repl