+++ packages/pglite/scripts/bundle-wasm.ts
@@ -76,6 +76,12 @@ async function main() {
     '.js',
     '.cjs',
   ])
+  await findAndReplaceInDir('./dist/age', /\.\.\/release\//g, '', [
+    '.js',
+    '.cjs',
+  ])
+  await findAndReplaceInDir('./dist/pg_textsearch', /\.\.\/release\//g, 
+    '', ['.js', '.cjs'])  
   await findAndReplaceInDir(
     './dist',
     `require("./postgres.js")`,