+++ packages/pglite/src/interface.ts
@@ -35,6 +35,11 @@ export interface ExecProtocolOptions {
   onNotice?: (notice: NoticeMessage) => void
 }
 
+export interface ExecProtocolOptionsStream {
+  syncToFs?: boolean
+  onRawData: (data: Uint8Array) => void
+}
+
 export interface ExtensionSetupResult<TNamespace = any> {
   emscriptenOpts?: any
   namespaceObj?: TNamespace
@@ -92,6 +97,7 @@ export interface PGliteOptions<TExtensions extends Extensions = Extensions> {
   fsBundle?: Blob | File
   parsers?: ParserOptions
   serializers?: SerializerOptions
+  startParams?: string[]
 }
 
 export type PGliteInterface<T extends Extensions = Extensions> =
@@ -118,6 +124,10 @@ export type PGliteInterface<T extends Extensions = Extensions> =
       message: Uint8Array,
       options?: ExecProtocolOptions,
     ): Promise<Uint8Array>
+    execProtocolRawStream(
+      message: Uint8Array,
+      options?: ExecProtocolOptionsStream,
+    ): Promise<void>
     execProtocol(
       message: Uint8Array,
       options?: ExecProtocolOptions,