+++ packages/pglite-repl/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @electric-sql/pglite-repl
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [37fb39e]
+  - @electric-sql/pglite@0.4.1
+  - @electric-sql/pglite-react@0.3.1
+
+## 0.3.0
+
+### Minor Changes
+
+- Updated dependencies [d848955]
+  - @electric-sql/pglite-react@0.3.0
+  - @electric-sql/pglite@0.4.0
+
+## 0.2.34
+
+### Patch Changes
+
+- Updated dependencies [3dfa40f]
+  - @electric-sql/pglite@0.3.16
+  - @electric-sql/pglite-react@0.2.34
+
 ## 0.2.33
 
 ### Patch Changes