+++ packages/pglite-socket/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @electric-sql/pglite-socket
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [37fb39e]
+  - @electric-sql/pglite@0.4.1
+
+## 0.1.0
+
+### Minor Changes
+
+- Updated dependencies [d848955]
+  - @electric-sql/pglite-postgis@0.0.1
+  - @electric-sql/pglite@0.4.0
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [3dfa40f]
+  - @electric-sql/pglite@0.3.16
+
 ## 0.0.21
 
 ### Patch Changes